refactor(2025-12-roundabouts): move BarChart component to lib folder
