+++ packages/twreporter/ddd/2025-12-roundabouts/src/lib/components/BarChart.svelte
